+++ src/main/java/com/amro/recipes/dao/model/RecipeIngredient.java
@@ -1,6 +1,7 @@
 package com.amro.recipes.dao.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.sun.xml.bind.v2.TODO;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -17,7 +18,7 @@ public class RecipeIngredient implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, unique = true)
-    private Integer id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "fk_ingredient_id", referencedColumnName = "id")
@@ -27,6 +28,13 @@ public class RecipeIngredient implements Serializable {
     @JoinColumn(name = "fk_recipe_id", referencedColumnName = "id")
     private Recipe rfRecipes;
 
+//    @Column(name = "amount", nullable = false)
+//    private float amount;
+
+//    @Column(name = "serving_number", nullable = false)
+//    private Integer servingNumber;
+
+    // TODO: 09/11/2022 create composite index
 }
